move kernel/fs/ext2 to libs/ext2; remove clone an usize; remove Deref,DerefMut implement
