+++ kernel/fs/mod.rs
@@ -1,5 +1,4 @@
 pub mod devfs;
-pub mod ext2;
 pub mod file_system;
 pub mod initramfs;
 pub mod inode;
